+++ packages/example/src/pages/TabDemo/index.ts
@@ -1,0 +1,2 @@
+import { TabDemo } from './TabDemo'
+export default TabDemo