+++ packages/table/src/TableExcel.tsx
@@ -121,7 +121,7 @@ export const TableExcel = forwardRef<AntdTableRef, AntdTableProps>((props, table
   })
 
   /** tbody单元格合并功能 */
-  useMergeCells({ showBorder: true })
+  useMergeCells({ showBorder: true, containerSize })
 
   /** tbody区域选中功能 + 拖拽边缘自动滚动 */
   useAreaSelected({