+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @grid-table/basic
 
+## 0.6.54
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.54
+
+## 0.6.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.53
+
 ## 0.6.49
 
 ### Patch Changes