+++ packages/example/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @grid-table/example
 
+## 0.0.64
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.54
+  - @grid-table/basic@0.6.54
+  - @grid-table/view@0.6.54
+  - @grid-table/excel@0.5.54
+  - @grid-table/pivot@0.5.57
+
+## 0.0.63
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.53
+  - @grid-table/basic@0.6.53
+  - @grid-table/view@0.6.53
+  - @grid-table/excel@0.5.53
+  - @grid-table/pivot@0.5.56
+
 ## 0.0.62
 
 ### Patch Changes