+++ packages/excel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @grid-table/excel
 
+## 0.5.54
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.54
+  - @grid-table/basic@0.6.54
+  - @grid-table/view@0.6.54
+
+## 0.5.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/core@0.6.53
+  - @grid-table/basic@0.6.53
+  - @grid-table/view@0.6.53
+
 ## 0.5.52
 
 ### Patch Changes